+++ src/components/bottomTabBar/ItemDetailsTabBar.tsx
@@ -1,10 +1,14 @@
 import { View } from "react-native";
 import Button from "../Button";
 import { bottomTabBar } from "@/styles/components/bottomTabBar";
+import DateSelector from "../itemList/DateSelector";
 
 const ItemDetailsTabBar = () => {
     const handlePress = () => {
-
+        console.log("asdf");
+        return (
+            <DateSelector />
+        )
     }
 
     return (
